% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{comparativeMethod_qPCR.analysis}
\alias{comparativeMethod_qPCR.analysis}
\title{qPCR comparative Method}
\usage{
comparativeMethod_qPCR.analysis(Gene, Ref)
}
\arguments{
\item{Gene}{delta CT value of the gene}

\item{Ref}{delta CT value of the reference gene}
}
\value{
a \code{data.frame} with delta-delta CT values
}
\description{
perform the comparative method for the qPCR analysis
}
\author{
Claus Weinholdt
}
