% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{comparativeMethod_qPCR.EGF.log2FC}
\alias{comparativeMethod_qPCR.EGF.log2FC}
\title{log2FC of EGF treatment for qPCR comparative Method}
\usage{
comparativeMethod_qPCR.EGF.log2FC(comparativeMethod_qPCR, CL, Treat, nREP = 4)
}
\arguments{
\item{comparativeMethod_qPCR}{generated by function comparativeMethod_qPCR.analysis()}

\item{Ref}{delta CT value of the reference gene}
}
\value{
a \code{data.frame} with log2-fold changes; 'ddCT.21$ddCT.logFC.21' are the log2FC for the EGF treatment.
}
\description{
calculates log2FC for the comparative method for the qPCR analysis based on the EGF treatment
}
\author{
Claus Weinholdt
}
