% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{comparativeMethod_qPCR.RNAi.log2FC}
\alias{comparativeMethod_qPCR.RNAi.log2FC}
\title{log2FC of RNAi for qPCR comparative Method}
\usage{
comparativeMethod_qPCR.RNAi.log2FC(comparativeMethod_qPCR)
}
\arguments{
\item{comparativeMethod_qPCR}{generated by function comparativeMethod_qPCR.analysis()}

\item{Ref}{delta CT value of the reference gene}
}
\value{
a \code{data.frame} with log2-fold changes; 'ddCT.C1C0$ddCT.logFC.C1C0' are the log2FC for class c
}
\description{
calculates log2FC for the comparative method for the qPCR analysis based on the RNAi experimental design
}
\author{
Claus Weinholdt
}
