% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExpData}
\alias{ExpData}
\title{Expression data set of XXX}
\format{lists : "source" ,"E"   ,    "genes" ,  "other" ,  "targets"}
\usage{
ExpData
}
\description{
A dataset containing the EListRaw objekt with
}
\keyword{datasets}
