% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{f.input.list}
\alias{f.input.list}
\title{Venndiagramm Wrapper}
\usage{
f.input.list(inList, VennPlot = TRUE, VennOut = FALSE)
}
\arguments{
\item{inList}{is list}

\item{VennPlot}{is boolean to enable venn plot}

\item{VennOut}{is boolean to enable return of venn-plot object}
}
\description{
Venndiagramm for list
}
\author{
Claus Weinholdt
}
