% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{get.gene.classes}
\alias{get.gene.classes}
\title{get Results of Posterior data}
\usage{
get.gene.classes(data, indexing="max",filter=0.75, DoPlot=FALSE)
}
\arguments{
\item{data}{is Posterior matrix}

\item{indexing}{is the indexing method "max" or "min"}

\item{filter}{for Posterior}

\item{DoPlot}{if TRUE plot histogram of Posterior}
}
\value{
a \code{list} of assigned genes
}
\description{
get the Genes assigned to the best class
}
\author{
Claus Weinholdt
}
