% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{get.IC}
\alias{get.IC}
\title{get InformationCriterion}
\usage{
get.IC(L,npar,k,IC)
}
\arguments{
\item{L}{is log liklelihood matrix}

\item{npar}{represents the number of parameters in the fitted model}

\item{k}{is number of ovservations}

\item{IC}{is BIC or AIC}
}
\value{
a \code{list}
}
\description{
get InformationCriterion (AIC or BIC) of LogLik matrix
}
\author{
Claus Weinholdt
}
