% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{get.log2Mean.and.log2FC}
\alias{get.log2Mean.and.log2FC}
\title{normalize ExpData}
\usage{
get.log2Mean.and.log2FC(normData)
}
\arguments{
\item{normData}{is the normData data object}
}
\value{
a \code{list} with mean , std.err , log2FC and std.err of log2FC
}
\description{
.AAA
}
\note{
\code{std.error.xy = sqrt( std.error(x)^2 +  std.error(y)^2)}
}
\author{
Claus Weinholdt
}
