% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{get.Posterior}
\alias{get.Posterior}
\title{get Posterior from BIC}
\usage{
get.Posterior(B,Pis)
}
\arguments{
\item{B}{is BIC matrix}

\item{Pis}{class prior}
}
\value{
a \code{matrix} of  Posterior
}
\description{
get Posterior from BIC
}
\author{
Claus Weinholdt
}
