% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{getQPCR}
\alias{getQPCR}
\title{getQPCR}
\usage{
getQPCR(set = "SF767", expressionOut = FALSE)
}
\arguments{
\item{set}{define which data set is used; defalut 'SF767'}

\item{expressionOut}{if TRUE return of log2 exprssion}
}
\value{
a \code{data.table}
}
\description{
get qPCR
}
\author{
Claus Weinholdt
}
