% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Illumina_to_REFSEQ_MRNA}
\alias{Illumina_to_REFSEQ_MRNA}
\title{Illumina to REFSEQ_MRNA tabel
GEOD.adf <-  fread('Illumina HumanHT-12_V4.0_A-GEOD-13475.adf.txt')
setnames(GEOD.adf,'Reporter Database Entry [genbank]','genbank')
setnames(GEOD.adf,'Reporter Name','illumina_humanht_12_v4')
REFSEQ_MRNA <- sapply(GEOD.adf[['genbank']], function(x) stringr::str_split(x,'\\.') )
GEOD.adf$REFSEQ_MRNA <- sapply(REFSEQ_MRNA, "[[", 1 ) 
setkey(GEOD.adf,'illumina_humanht_12_v4')
Illumina_to_REFSEQ_MRNA <- GEOD.adf
devtools::use_data(Illumina_to_REFSEQ_MRNA)
A dataset containing the}
\format{A data frame with 47323 rows and 5 variables:}
\usage{
Illumina_to_REFSEQ_MRNA
}
\description{
Illumina to REFSEQ_MRNA tabel
GEOD.adf <-  fread('Illumina HumanHT-12_V4.0_A-GEOD-13475.adf.txt')
setnames(GEOD.adf,'Reporter Database Entry [genbank]','genbank')
setnames(GEOD.adf,'Reporter Name','illumina_humanht_12_v4')
REFSEQ_MRNA <- sapply(GEOD.adf[['genbank']], function(x) stringr::str_split(x,'\\.') )
GEOD.adf$REFSEQ_MRNA <- sapply(REFSEQ_MRNA, "[[", 1 ) 
setkey(GEOD.adf,'illumina_humanht_12_v4')
Illumina_to_REFSEQ_MRNA <- GEOD.adf
devtools::use_data(Illumina_to_REFSEQ_MRNA)
A dataset containing the
}
\keyword{datasets}
