% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{InformationCriterion}
\alias{InformationCriterion}
\title{calucate BIC and AIC}
\usage{
lInformationCriterion(logL,npar,k , IC='BIC')
}
\arguments{
\item{logL}{is log liklelihood}

\item{npar}{represents the number of parameters in the fitted model}

\item{k}{is number of ovservations}

\item{IC}{is BIC or AIC}
}
\value{
a \code{value}
}
\description{
calucate BIC and AIC
}
\author{
Claus Weinholdt
}
