% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{intersect3}
\alias{intersect3}
\title{intersect 3-Set}
\usage{
intersect3(l1, l2, l3)
}
\arguments{
\item{l1}{is the 1. set}

\item{l2}{is the 2. set}

\item{l3}{is the 3. set}
}
\description{
intersect for 3-Set
}
\author{
Claus Weinholdt

Claus Weinholdt
}
