% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{intersect5}
\alias{intersect5}
\title{intersect 5-Set}
\usage{
intersect5(l1, l2, l3, l4, l5)
}
\arguments{
\item{l1}{is the 1. set}

\item{l2}{is the 2. set}

\item{l3}{is the 3. set}

\item{l4}{is the 4. set}

\item{l5}{is the 5. set}
}
\description{
intersect for 5-Set
}
\author{
Claus Weinholdt

Claus Weinholdt
}
