% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{logLikelihoodOfnormData}
\alias{logLikelihoodOfnormData}
\title{calucate the the log likelihood for each class for each gene}
\usage{
logLikelihoodOfnormData(x)
}
\arguments{
\item{x}{is maxtrix with normalized expression e.g. normData$E}
}
\value{
a \code{matrix} with log likelihoods
}
\description{
calucate the the log likelihood for each class for each gene
}
\author{
Claus Weinholdt
}
