% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{make.Enrichment.data}
\alias{make.Enrichment.data}
\title{make data for Enrichment analysis}
\usage{
make.Enrichment.data(data, LsetForFC, normData)
}
\arguments{
\item{data}{is Posterior matrix}

\item{LsetForFC}{is the Lset of the data. needed for the logFC calculation}

\item{normData}{is the normData object}
}
\value{
a \code{list} with logFC and Symbols
}
\description{
make data for Enrichment analysis
}
\author{
Claus Weinholdt
}
