% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{make.plot.data.FC.Ill.qPCR}
\alias{make.plot.data.FC.Ill.qPCR}
\title{make.plot.data.FC.Ill.qPCR}
\usage{
make.plot.data.FC.Ill.qPCR(qCPRdata, MeanFoldChangeClass, class = "c")
}
\arguments{
\item{qCPRdata}{is data.table with qPCR data}

\item{MeanFoldChangeClass}{is data.table with Illumina data}

\item{class}{for the plots}
}
\value{
a \code{data.frame} as plot input
}
\description{
make.plot.data.FC.Ill.qPCR
}
\author{
Claus Weinholdt
}
