% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{normalizeExpData}
\alias{normalizeExpData}
\title{normalize ExpData}
\usage{
normalizeExpData(DetectionPval = 0.05 , DetectionPvalNumber = "ALL")
}
\arguments{
\item{DetectionPval}{is the dectecion p value ot the Illumina BeatChip}

\item{DetectionPvalNumber}{miniml number of samples with DetectionPval. If DetectionPvalNumber is "ALL" then DetectionPvalNumber equal to the number of samples}
}
\value{
a \code{limma object}
}
\description{
.AAA
}
\author{
Claus Weinholdt
}
