% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qPCR_SF767}
\alias{qPCR_SF767}
\title{Delta CT values of RNAi treatment for SF767}
\format{A data frame with 18 rows and 10 variables:
\describe{
  \item{CellLine}{Cell line ID}
  \item{Treatment}{Treatment}
  \item{GAPDH}{deltaCt of GAPDH}
  \item{CKAP2L}{deltaCt of CKAP2L}
  \item{ROCK1}{deltaCt of ROCK1}
  \item{TPR}{deltaCt of TPR}
  \item{ALDH4A1}{deltaCt of ALDH4A1}
  \item{GALNS}{deltaCt of GALNS}
  \item{CLCA2}{deltaCt of CLCA2}
}}
\usage{
qPCR_SF767
}
\description{
A dataset containing the
}
\keyword{datasets}
