% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qPCR_SF767_LNZ308}
\alias{qPCR_SF767_LNZ308}
\title{Delta CT values of EGF treatment for SF767 and LNZ308}
\format{A data frame with 12 rows and 9 variables:
\describe{
  \item{ID}{qPCR ID}
  \item{CellLine}{Cell line ID}
  \item{Treatment}{Treatment}
  \item{GAPDH}{deltaCt of GAPDH}
  \item{CKAP2L}{deltaCt of CKAP2L}
  \item{ROCK1}{deltaCt of ROCK1}
  \item{TPR}{deltaCt of TPR}
  \item{ALDH4A1}{deltaCt of ALDH4A1}
  \item{GALNS}{deltaCt of GALNS}
}}
\usage{
qPCR_SF767_LNZ308
}
\description{
A dataset containing the
}
\keyword{datasets}
