% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RawQPCRsf}
\alias{RawQPCRsf}
\title{Expression data set of all qPCRs}
\format{A data frame with 30 rows and 14 variables:
\describe{
  \item{Probe}{Name of celline}
  \item{NR}{Sample number}
  \item{CKAP2L_ctGen.ctRef}{deltaCt of CKAP2L}
  \item{CKAP2L_relative.Werte}{relativ expression of CKAP2L}
  ...
  \item{GALNSctGen.ctRef}{deltaCt of GALNS}
  \item{GALNSrelative.Werte}{relativ expression of GALNS}
}}
\usage{
RawQPCRsf
}
\description{
A dataset containing the
}
\keyword{datasets}
