% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{thememapBarplot}
\alias{thememapBarplot}
\title{ggplot thememap for barplot}
\usage{
thememapBarplot(base_size = 12, legend_key_size = 0.4, base_family = "",
  colgridmajor = "grey70")
}
\arguments{
\item{base_size}{size of font}

\item{legend_key_size}{size of key for the legend}

\item{base_family}{base family}

\item{colgridmajor}{color of major grid}
}
\value{
a \code{theme}
}
\description{
ggplot thememap for barplot
}
\author{
Claus Weinholdt
}
