% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{union3}
\alias{union3}
\title{union 3-Set}
\usage{
union3(l1, l2, l3)
}
\arguments{
\item{l1}{is the 1. set}

\item{l2}{is the 2. set}

\item{l3}{is the 3. set}
}
\description{
union for 3-Set
}
\author{
Claus Weinholdt

Claus Weinholdt
}
