% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennFunction.R
\name{union4}
\alias{union4}
\title{union 4-Set}
\usage{
union4(l1, l2, l3, l4)
}
\arguments{
\item{l1}{is the 1. set}

\item{l2}{is the 2. set}

\item{l3}{is the 3. set}

\item{l4}{is the 4. set}
}
\description{
union for 4-Set
}
\author{
Claus Weinholdt

Claus Weinholdt
}
