% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{Density.NV.fit.plot}
\alias{Density.NV.fit.plot}
\title{Density plot for NV fit}
\usage{
Density.NV.fit.plot(id, normData, ALL.MUs, ALL.VARs, useGroup = NA,
  DOplot = FALSE, basesize = 14, GrAblack = TRUE, onlySYMBOL = FALSE)
}
\arguments{
\item{id}{is a Illuminan id}

\item{normData}{is the normData data object}

\item{ALL.MUs}{matrix with all means from logLikelihoodOfnormData()}

\item{ALL.VARs}{matrix with all variances from logLikelihoodOfnormData()}

\item{useGroup}{if set to a group (eg. "a","b","c" or "d") the row is colored}

\item{DOplot}{if TRUE plot is printed}

\item{basesize}{size of font}

\item{GrAblack}{if TRUE coloring group 'a' in black because we do not use a Indicator variable}

\item{onlySYMBOL}{if TRUE title is only gene SYMBOL}
}
\value{
a \code{list} with mean , std.err , log2FC and std.err of log2FC
}
\description{
Density plot for NV fit
}
\author{
Claus Weinholdt
}
