% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{make.enrichment.data}
\alias{make.enrichment.data}
\title{make data for enrichment analysis}
\usage{
make.enrichment.data(data, LsetForFC, normData)
}
\arguments{
\item{data}{is Posterior matrix}

\item{LsetForFC}{is the Lset of the data. needed for the logFC calculation}

\item{normData}{is the normData object}
}
\value{
a \code{list} with logFC and Symbols
}
\description{
make data for enrichment analysis
}
\author{
Claus Weinholdt
}
