% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSC.R
\name{normalizeExpData}
\alias{normalizeExpData}
\title{normalize ExpData}
\usage{
normalizeExoData()
}
\arguments{
\item{pval}{is the dectecion p value ot the Illumina BeatChip}
}
\value{
a \code{limma object}
}
\description{
.AAA
}
\author{
Claus Weinholdt
}
